+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.6.0] - 2025-06-xx
+
+- New slide parser logic to support more complex front matter slide structures
+- [#156](https://github.com/estruyf/vscode-demo-time/issues/156): Add preview action to hide the
+  mouse cursor and action bar
+- [#162](https://github.com/estruyf/vscode-demo-time/issues/162): Added a laser pointer action to
+  the slide view
+
 ## [1.5.1] - 2025-06-17
 
 - fix: add 'image' to allowed frontmatter properties in SlideParser